third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-style-computed-expected.txt

This is a testharness.js-based test.
PASS Property font-style value 'normal' computes to 'normal'
PASS Property font-style value 'italic' computes to 'italic'
FAIL Property font-style value 'oblique' computes to 'oblique' assert_equals: expected "oblique" but got "italic"
Harness: the test ran to completion.
